@@ -20,16 +20,17 @@
       "psr-4": { "Lhm\\Tests\\": "tests/" }
     },
     "require": {
-      "php": "^5.6 || ^7.0",
+      "php": ">=7.2",
       "psr/log": "^1.0",
-      "symfony/console": "^2.0 || ^3.0",
-      "symfony/event-dispatcher": "^2.0 || ^3.0",
-      "symfony/monolog-bridge": "^2.0 || ^3.0"
+      "symfony/console": "^3.4|^4.0|^5.0",
+      "symfony/event-dispatcher": "^3.4|^4.0|^5.0",
+      "symfony/monolog-bridge": "^3.4|^4.0|^5.0",
+      "ext-pdo": "*"
     },
     "require-dev": {
-        "monolog/monolog": "~1.0",
-        "phpunit/phpunit": "^5.7 || ^6.1",
-        "robmorgan/phinx": "~0.9"
+        "monolog/monolog": "^2",
+        "phpunit/phpunit": "^9",
+        "robmorgan/phinx": "^0.12"
     },
     "suggest": {
         "robmorgan/phinx": "Advanced database migration tool",
