@@ -26,7 +26,7 @@
 /** @var callable $autoloader */
 $autoloader = require __DIR__ . '/../src/composer_autoloader.php';
 
-if (!$autoloader()) {
+if ( ! $autoloader()) {
     die(
         'You need to set up the project dependencies using the following commands:' . PHP_EOL .
         'curl -s http://getcomposer.org/installer | php' . PHP_EOL .
