@@ -5,7 +5,6 @@ namespace Lhm\Console\Command;
 
 
 use Lhm\Lhm;
-use Monolog\Handler\StreamHandler;
 use Monolog\Logger;
 use Phinx\Console\Command\AbstractCommand;
 use Symfony\Bridge\Monolog\Handler\ConsoleHandler;
@@ -24,11 +23,13 @@ class Cleanup extends AbstractCommand
 
         $this->setName('cleanup')
             ->setDescription('Cleanup LHM tables, old archives and triggers')
-            ->setHelp(sprintf(
-                '%sCleanup LHM tables, old archives and triggers. Defaults to a dry-run unless --run is specified.%s',
-                PHP_EOL,
-                PHP_EOL
-            ));
+            ->setHelp(
+                sprintf(
+                    '%sCleanup LHM tables, old archives and triggers. Defaults to a dry-run unless --run is specified.%s',
+                    PHP_EOL,
+                    PHP_EOL
+                )
+            );
 
         $this->addOption('--environment', '-e', InputOption::VALUE_REQUIRED, 'The target environment');
 
@@ -36,25 +37,31 @@ class Cleanup extends AbstractCommand
         $this->addOption('--run', 'r', InputOption::VALUE_NONE, 'Apply the cleanup operations.');
 
         // Archives older than the specified UTC date + time will be dropped.
-        $this->addOption('--until', 'u', InputOption::VALUE_REQUIRED, 'Drop archive tables older than the specified date at UTC (YYYY-MM-DD_hh:mm:ss).');
+        $this->addOption(
+            '--until',
+            'u',
+            InputOption::VALUE_REQUIRED,
+            'Drop archive tables older than the specified date at UTC (YYYY-MM-DD_hh:mm:ss).'
+        );
     }
 
     /**
      * Cleanup the database
      *
-     * @param InputInterface $input
+     * @param InputInterface  $input
      * @param OutputInterface $output
-     * @throws \RuntimeException
-     * @throws \InvalidArgumentException
+     *
      * @return void
+     * @throws \InvalidArgumentException
+     * @throws \RuntimeException
      */
     protected function execute(InputInterface $input, OutputInterface $output)
     {
-        if (!$this->getConfig()) {
+        if ( ! $this->getConfig()) {
             $this->loadConfig($input, $output);
         }
 
-        $this->loadManager($output);
+        $this->loadManager($input, $output);
 
         $environment = $input->getOption('environment');
 
@@ -74,7 +81,7 @@ class Cleanup extends AbstractCommand
             $output->writeln('<info>using database</info> ' . $envOptions['name']);
         }
 
-        $run = (bool)$input->getOption('run');
+        $run   = (bool)$input->getOption('run');
         $until = $input->getOption('until');
 
         if ($until) {
@@ -95,7 +102,9 @@ class Cleanup extends AbstractCommand
         $options = [];
         if ($until) {
             $options['until'] = $until;
-            $output->writeln('<warning>LHM will drop archives created before ' . $until->format('Y-M-D H:i:s T') . '</warning>');
+            $output->writeln(
+                '<warning>LHM will drop archives created before ' . $until->format('Y-M-D H:i:s T') . '</warning>'
+            );
         }
 
 
