@@ -3,7 +3,9 @@
 namespace Lhm;
 
 use Phinx\Db\Adapter\AdapterInterface;
+use Phinx\Db\Table as PhinxTable;
 use Psr\Log\LoggerInterface;
+use RuntimeException;
 
 
 class Invoker
@@ -12,7 +14,7 @@ class Invoker
     const LOCK_WAIT_TIMEOUT_DELTA = -2;
 
     /**
-     * @var Table
+     * @var PhinxTable
      */
     protected $origin;
 
@@ -36,40 +38,24 @@ class Invoker
 
     /**
      * @param AdapterInterface $adapter
-     * @param Table $origin
-     * @param array $options
-     *                      - `stride` integer
+     * @param PhinxTable       $origin
+     * @param array            $options
+     *                          - `stride` integer
      *                          Size of a chunk (defaults to 2000)
-     *                      - `atomic_switch` boolean
+     *                          - `atomic_switch` boolean
      *                          Enable atomic switching (defaults to true)
-     *                      - `retry_sleep_time` integer
+     *                          - `retry_sleep_time` integer
      *                          How long should the switch wait until retrying ( defaults to 10 )
-     *                      - `max_retries` integer
+     *                          - `max_retries` integer
      *                          How many times the switch should be attempted ( defaults to 600 )
-     *                      - `archive_name` string
+     *                          - `archive_name` string
      *                          Name of the archive table ( defaults to 'lhma_' . gmdate('Y_m_d_H_i_s') . "_{$origin->getName()}" )
      */
-    public function __construct(AdapterInterface $adapter, \Phinx\Db\Table $origin, array $options = [])
+    public function __construct(AdapterInterface $adapter, PhinxTable $origin, array $options = [])
     {
         $this->options = $options;
         $this->adapter = $adapter;
-        $this->origin = $origin;
-    }
-
-    /**
-     * @param LoggerInterface $logger
-     */
-    public function setLogger(LoggerInterface $logger)
-    {
-        $this->logger = $logger;
-    }
-
-    /**
-     * @return LoggerInterface
-     */
-    protected function getLogger()
-    {
-        return $this->logger ?: new NullLogger();
+        $this->origin  = $origin;
     }
 
     /**
@@ -89,18 +75,18 @@ class Invoker
 
         $sqlHelper = new SqlHelper($this->adapter);
 
-        if (!isset($options['atomic_switch'])) {
-
+        if ( ! isset($options['atomic_switch'])) {
             if ($sqlHelper->supportsAtomicSwitch()) {
                 $this->options['atomic_switch'] = true;
             } else {
                 $version = $sqlHelper->versionString();
-                throw new \RuntimeException("Using mysql {$version}. You must explicitly set `options['atomic_switch']` (re SqlHelper::supportsAtomicSwitch)");
+                throw new RuntimeException(
+                    "Using mysql {$version}. You must explicitly set `options['atomic_switch']` (re SqlHelper::supportsAtomicSwitch)"
+                );
             }
-
         }
 
-        if (!$this->destination) {
+        if ( ! $this->destination) {
             $this->destination = $this->temporaryTable();
         }
 
@@ -123,10 +109,60 @@ class Invoker
 
         $migration($this->destination);
 
-        $entangler->run(function () use ($chunker, $switcher) {
-            $chunker->run();
-            $switcher->run();
-        });
+        $entangler->run(
+            function () use ($chunker, $switcher) {
+                $chunker->run();
+                $switcher->run();
+            }
+        );
+    }
+
+    /**
+     * @return Table
+     * @throws RuntimeException
+     */
+    public function temporaryTable()
+    {
+        if ($this->destination) {
+            return $this->destination;
+        }
+
+        $temporaryTableName = $this->temporaryTableName();
+
+        if ($this->adapter->hasTable($temporaryTableName)) {
+            throw new RuntimeException("The table `{$temporaryTableName}` already exists.");
+        }
+
+        $this->getLogger()->info("Creating temporary table `{$temporaryTableName}` from `{$this->origin->getName()}`");
+        $this->adapter->query("CREATE TABLE {$temporaryTableName} LIKE {$this->origin->getName()}");
+
+        return new Table($temporaryTableName, [], $this->adapter);
+    }
+
+    /**
+     * Returns the temporary table name.
+     *
+     * @return string
+     */
+    public function temporaryTableName()
+    {
+        return "lhmn_{$this->origin->getName()}";
+    }
+
+    /**
+     * @return LoggerInterface
+     */
+    protected function getLogger()
+    {
+        return $this->logger ?: new NullLogger();
+    }
+
+    /**
+     * @param LoggerInterface $logger
+     */
+    public function setLogger(LoggerInterface $logger)
+    {
+        $this->logger = $logger;
     }
 
     protected function setSessionLockWaitTimeouts()
@@ -135,8 +171,12 @@ class Invoker
         $logger->debug("Getting mysql session lock wait timeouts");
 
         //TODO File a bug with Phinx. $adapter->query does not return an array ( returns a PDOStatement )
-        $globalInnodbLockWaitTimeout = $this->adapter->query("SHOW GLOBAL VARIABLES LIKE 'innodb_lock_wait_timeout'")->fetchColumn(0);
-        $globalLockWaitTimeout = $this->adapter->query("SHOW GLOBAL VARIABLES LIKE 'lock_wait_timeout'")->fetchColumn(0);
+        $globalInnodbLockWaitTimeout = $this->adapter->query(
+            "SHOW GLOBAL VARIABLES LIKE 'innodb_lock_wait_timeout'"
+        )->fetchColumn(0);
+        $globalLockWaitTimeout       = $this->adapter->query(
+            "SHOW GLOBAL VARIABLES LIKE 'lock_wait_timeout'"
+        )->fetchColumn(0);
 
 
         if ($globalInnodbLockWaitTimeout) {
@@ -154,38 +194,5 @@ class Invoker
 
             $this->adapter->query("SET SESSION lock_wait_timeout={$value}");
         }
-    }
-
-
-    /**
-     * @return Table
-     * @throws \RuntimeException
-     */
-    public function temporaryTable()
-    {
-
-        if ($this->destination) {
-            return $this->destination;
-        }
-
-        $temporaryTableName = $this->temporaryTableName();
-
-        if ($this->adapter->hasTable($temporaryTableName)) {
-            throw new \RuntimeException("The table `{$temporaryTableName}` already exists.");
-        }
-
-        $this->getLogger()->info("Creating temporary table `{$temporaryTableName}` from `{$this->origin->getName()}`");
-        $this->adapter->query("CREATE TABLE {$temporaryTableName} LIKE {$this->origin->getName()}");
-
-        return new \Lhm\Table($temporaryTableName, [], $this->adapter);
-    }
-
-    /**
-     * Returns the temporary table name.
-     * @return string
-     */
-    public function temporaryTableName()
-    {
-        return "lhmn_{$this->origin->getName()}";
     }
 }
