@@ -5,6 +5,8 @@ namespace Lhm;
 
 
 use Phinx\Db\Adapter\AdapterInterface;
+use Phinx\Db\Table as PhinxTable;
+use RuntimeException;
 
 
 class LockedSwitcher extends Command
@@ -12,11 +14,12 @@ class LockedSwitcher extends Command
     /** @var AdapterInterface */
     protected $adapter;
 
-    /** @var \Phinx\Db\Table */
+    /** @var PhinxTable */
     protected $origin;
 
-    /** @var \Lhm\Table */
+    /** @var Table */
     protected $destination;
+    
     /**
      * @var SqlHelper
      */
@@ -24,24 +27,60 @@ class LockedSwitcher extends Command
 
     /**
      * @param AdapterInterface $adapter
-     * @param \Phinx\Db\Table $origin
-     * @param \Lhm\Table $destination
-     * @param array $options Table options:
-     *                      - `retry_sleep_time`
-     *                      - `max_retries`
-     *                      - `archive_name`
+     * @param PhinxTable       $origin
+     * @param Table            $destination
+     * @param array            $options Table options:
+     *                                  - `retry_sleep_time`
+     *                                  - `max_retries`
+     *                                  - `archive_name`
      */
-    public function __construct(AdapterInterface $adapter, \Phinx\Db\Table $origin, \Lhm\Table $destination, array $options = [])
-    {
+    public function __construct(
+        AdapterInterface $adapter,
+        PhinxTable $origin,
+        Table $destination,
+        array $options = []
+    ) {
         $this->options = $options + [
                 'retry_sleep_time' => 10,
-                'max_retries' => 600,
-                'archive_name' => 'lhma_' . gmdate('Y_m_d_H_i_s') . "_{$origin->getName()}"
+                'max_retries'      => 600,
+                'archive_name'     => 'lhma_' . gmdate('Y_m_d_H_i_s') . "_{$origin->getName()}",
             ];
 
-        $this->adapter = $adapter;
-        $this->origin = $origin;
+        $this->adapter     = $adapter;
+        $this->origin      = $origin;
         $this->destination = $destination;
+    }
+
+    /**
+     * @return array
+     */
+    public function getOptions()
+    {
+        return $this->options;
+    }
+
+    /**
+     * @throws RuntimeException
+     */
+    protected function validate()
+    {
+        if ($this->adapter->hasTable($this->origin->getName()) && $this->adapter->hasTable(
+                $this->destination->getName()
+            )) {
+            return;
+        }
+
+        throw new RuntimeException(
+            "Table `{$this->origin->getName()}` and `{$this->destination->getName()}` must exist."
+        );
+    }
+
+    /**
+     * Revert the switch
+     */
+    protected function revert()
+    {
+        $this->adapter->query($this->getSqlHelper()->tagged('unlock tables'));
     }
 
     /**
@@ -49,7 +88,7 @@ class LockedSwitcher extends Command
      */
     public function getSqlHelper()
     {
-        if (!$this->sqlHelper) {
+        if ( ! $this->sqlHelper) {
             $this->sqlHelper = new SqlHelper($this->adapter);
         }
 
@@ -65,11 +104,14 @@ class LockedSwitcher extends Command
     }
 
     /**
-     * @return array
+     * Execute the switch
      */
-    public function getOptions()
+    protected function execute()
     {
-        return $this->options;
+        $sqlHelper = $this->getSqlHelper();
+        foreach ($this->statements() as $statement) {
+            $this->adapter->query($sqlHelper->tagged($statement));
+        }
     }
 
     /**
@@ -81,33 +123,15 @@ class LockedSwitcher extends Command
     }
 
     /**
-     * @return array
-     */
-    protected function switchStatements()
-    {
-        $originTableName = $this->adapter->quoteTableName($this->origin->getName());
-        $destinationTableName = $this->adapter->quoteTableName($this->destination->getName());
-
-        $archiveName = $this->adapter->quoteTableName($this->options['archive_name']);
-
-        return [
-            "LOCK TABLE {$originTableName} write, {$destinationTableName} write",
-            "ALTER TABLE  {$originTableName} rename {$archiveName}",
-            "ALTER TABLE {$destinationTableName} rename {$originTableName}",
-            'COMMIT',
-            'UNLOCK TABLES'
-        ];
-    }
-
-    /**
      * @param array $switchStatements
+     *
      * @return array
      */
     protected function uncommitted($switchStatements)
     {
         $statements = [
             'set @lhm_auto_commit = @@session.autocommit',
-            'set session autocommit = 0'
+            'set session autocommit = 0',
         ];
 
         $statements = array_merge($statements, $switchStatements);
@@ -117,34 +141,22 @@ class LockedSwitcher extends Command
     }
 
     /**
-     * @throws \RuntimeException
+     * @return array
      */
-    protected function validate()
+    protected function switchStatements()
     {
-        if ($this->adapter->hasTable($this->origin->getName()) && $this->adapter->hasTable($this->destination->getName())) {
-            return;
-        }
+        $originTableName      = $this->adapter->quoteTableName($this->origin->getName());
+        $destinationTableName = $this->adapter->quoteTableName($this->destination->getName());
 
-        throw new \RuntimeException("Table `{$this->origin->getName()}` and `{$this->destination->getName()}` must exist.");
-    }
+        $archiveName = $this->adapter->quoteTableName($this->options['archive_name']);
 
-    /**
-     * Revert the switch
-     */
-    protected function revert()
-    {
-        $this->adapter->query($this->getSqlHelper()->tagged('unlock tables'));
-    }
-
-    /**
-     * Execute the switch
-     */
-    protected function execute()
-    {
-        $sqlHelper = $this->getSqlHelper();
-        foreach ($this->statements() as $statement) {
-            $this->adapter->query($sqlHelper->tagged($statement));
-        }
+        return [
+            "LOCK TABLE {$originTableName} write, {$destinationTableName} write",
+            "ALTER TABLE  {$originTableName} rename {$archiveName}",
+            "ALTER TABLE {$destinationTableName} rename {$originTableName}",
+            'COMMIT',
+            'UNLOCK TABLES',
+        ];
     }
 
 
