@@ -1,9 +1,12 @@
 <?php
+
 namespace Lhm\Tests\Integration;
 
 use Lhm\Chunker;
 use Lhm\SqlHelper;
+use Lhm\Table;
 use Phinx\Db\Adapter\AdapterInterface;
+use Phinx\Db\Table as PhinxTable;
 use Phinx\Db\Table\Column;
 use PHPUnit\Framework\TestCase;
 
@@ -17,11 +20,11 @@ class ChunkerTest extends TestCase
      */
     protected $adapter;
     /**
-     * @var \Phinx\Db\Table
+     * @var PhinxTable
      */
     protected $origin;
     /**
-     * @var \Lhm\Table
+     * @var Table
      */
     protected $destination;
 
@@ -30,43 +33,13 @@ class ChunkerTest extends TestCase
      */
     protected $sqlHelper;
 
-    protected function setUp()
-    {
-        parent::setUp();
-        $this->adapter = $this->getMockBuilder(AdapterInterface::class)->getMock();
-        $this->adapter
-            ->expects($this->any())
-            ->method('quoteColumnName')
-            ->will($this->returnCallback(function ($name) {
-                return "`{$name}`";
-            }));
-
-        $this->adapter
-            ->expects($this->any())
-            ->method('quoteTableName')
-            ->will($this->returnCallback(function ($name) {
-                return "'{$name}'";
-            }));
-
-        $this->origin = $this->getMockBuilder(\Phinx\Db\Table::class)->disableOriginalConstructor()->getMock();
-        $this->destination = $this->getMockBuilder(\Lhm\Table::class)->disableOriginalConstructor()->getMock();
-
-        $this->sqlHelper = new SqlHelper($this->adapter);
-    }
-
-    protected function tearDown()
-    {
-        unset($this->chunker, $this->adapter, $this->origin, $this->destination, $this->sqlHelper);
-        parent::tearDown();
-    }
-
     public function testRun()
     {
         /** @var Column[] $originColumns */
         $originColumns = [
             new Column(),
             new Column(),
-            new Column()
+            new Column(),
         ];
         $originColumns[0]->setName('id');
         $originColumns[1]->setName('name');
@@ -76,7 +49,7 @@ class ChunkerTest extends TestCase
         $destinationColumns = [
             new Column(),
             new Column(),
-            new Column()
+            new Column(),
         ];
         $destinationColumns[0]->setName('id');
         $destinationColumns[1]->setName('name');
@@ -111,67 +84,113 @@ class ChunkerTest extends TestCase
         $this->adapter
             ->expects($matcher)
             ->method('fetchRow')
-            ->will($this->returnCallback(function ($query) use ($matcher) {
-                switch ($matcher->getInvocationCount()) {
-                    case 1:
-                        $this->assertEquals(
-                            "SELECT MIN(`id`) FROM 'users'",
-                            $query
-                        );
-                        return [1];
+            ->will(
+                $this->returnCallback(
+                    function ($query) use ($matcher) {
+                        switch ($matcher->getInvocationCount()) {
+                            case 1:
+                                $this->assertEquals(
+                                    "SELECT MIN(`id`) FROM 'users'",
+                                    $query
+                                );
+                                return [1];
 
-                    case 2:
-                        $this->assertEquals(
-                            "SELECT MAX(`id`) FROM 'users'",
-                            $query
-                        );
-                        return [500];
-                    default:
+                            case 2:
+                                $this->assertEquals(
+                                    "SELECT MAX(`id`) FROM 'users'",
+                                    $query
+                                );
+                                return [500];
+                            default:
 
-                        return null;
-                        break;
-                }
-            }));
+                                return null;
+                                break;
+                        }
+                    }
+                )
+            );
 
         $matcher = $this->atLeastOnce();
         $this->adapter
             ->expects($matcher)
             ->method('query')
-            ->will($this->returnCallback(function ($query) use ($matcher) {
-                switch ($matcher->getInvocationCount()) {
-                    case 1:
-                        $this->assertEquals(
-                            "SELECT `COLUMN_NAME` FROM `information_schema`.`COLUMNS` WHERE (`TABLE_SCHEMA` = '') AND (`TABLE_NAME` = 'users') AND (`COLUMN_KEY` = 'PRI');",
-                            $query
-                        );
+            ->will(
+                $this->returnCallback(
+                    function ($query) use ($matcher) {
+                        switch ($matcher->getInvocationCount()) {
+                            case 1:
+                                $this->assertEquals(
+                                    "SELECT `COLUMN_NAME` FROM `information_schema`.`COLUMNS` WHERE (`TABLE_SCHEMA` = '') AND (`TABLE_NAME` = 'users') AND (`COLUMN_KEY` = 'PRI');",
+                                    $query
+                                );
 
-                        return 'id';
-                    case 2:
-                        $this->assertEquals(
-                            "INSERT IGNORE INTO 'users_new' (`id`,`name`) SELECT 'users'.`id`,'users'.`name` FROM 'users' WHERE 'users'.`id` BETWEEN 1 AND 200",
-                            $query
-                        );
-                        break;
-                    case 3:
-                        $this->assertEquals(
-                            "INSERT IGNORE INTO 'users_new' (`id`,`name`) SELECT 'users'.`id`,'users'.`name` FROM 'users' WHERE 'users'.`id` BETWEEN 201 AND 400",
-                            $query
-                        );
-                        break;
-                    case 4:
-                        $this->assertEquals(
-                            "INSERT IGNORE INTO 'users_new' (`id`,`name`) SELECT 'users'.`id`,'users'.`name` FROM 'users' WHERE 'users'.`id` BETWEEN 401 AND 500",
-                            $query
-                        );
-                        break;
-                    default:
-                        $this->fail('Unexpected query: ' . $query);
-                        break;
-                }
-            }));
+                                return 'id';
+                            case 2:
+                                $this->assertEquals(
+                                    "INSERT IGNORE INTO 'users_new' (`id`,`name`) SELECT 'users'.`id`,'users'.`name` FROM 'users' WHERE 'users'.`id` BETWEEN 1 AND 200",
+                                    $query
+                                );
+                                break;
+                            case 3:
+                                $this->assertEquals(
+                                    "INSERT IGNORE INTO 'users_new' (`id`,`name`) SELECT 'users'.`id`,'users'.`name` FROM 'users' WHERE 'users'.`id` BETWEEN 201 AND 400",
+                                    $query
+                                );
+                                break;
+                            case 4:
+                                $this->assertEquals(
+                                    "INSERT IGNORE INTO 'users_new' (`id`,`name`) SELECT 'users'.`id`,'users'.`name` FROM 'users' WHERE 'users'.`id` BETWEEN 401 AND 500",
+                                    $query
+                                );
+                                break;
+                            default:
+                                $this->fail('Unexpected query: ' . $query);
+                                break;
+                        }
+                    }
+                )
+            );
 
         $chunker = new Chunker($this->adapter, $this->origin, $this->destination, $this->sqlHelper, ['stride' => 200]);
         $chunker->run();
+    }
+
+    protected function setUp(): void
+    {
+        parent::setUp();
+        $this->adapter = $this->getMockBuilder(AdapterInterface::class)->getMock();
+        $this->adapter
+            ->expects($this->any())
+            ->method('quoteColumnName')
+            ->will(
+                $this->returnCallback(
+                    function ($name) {
+                        return "`{$name}`";
+                    }
+                )
+            );
+
+        $this->adapter
+            ->expects($this->any())
+            ->method('quoteTableName')
+            ->will(
+                $this->returnCallback(
+                    function ($name) {
+                        return "'{$name}'";
+                    }
+                )
+            );
+
+        $this->origin      = $this->getMockBuilder(PhinxTable::class)->disableOriginalConstructor()->getMock();
+        $this->destination = $this->getMockBuilder(Table::class)->disableOriginalConstructor()->getMock();
+
+        $this->sqlHelper = new SqlHelper($this->adapter);
+    }
+
+    protected function tearDown(): void
+    {
+        unset($this->chunker, $this->adapter, $this->origin, $this->destination, $this->sqlHelper);
+        parent::tearDown();
     }
 
 }
