@@ -1,9 +1,9 @@
 <?php
+
 namespace Lhm\Tests\Persistence;
 
 use Phinx\Db\Adapter\AdapterInterface;
 use Phinx\Db\Adapter\MysqlAdapter;
-use Symfony\Component\Console\Output\NullOutput;
 use PHPUnit\Framework\TestCase;
 
 abstract class AbstractPersistenceTest extends TestCase
@@ -11,7 +11,7 @@ abstract class AbstractPersistenceTest extends TestCase
     /** @var AdapterInterface */
     protected $adapter;
 
-    protected function setUp()
+    protected function setUp(): void
     {
         parent::setUp();
 
@@ -20,7 +20,7 @@ abstract class AbstractPersistenceTest extends TestCase
             'name' => getenv('LHM_DATABASE_NAME') ?: 'lhm_php_test',
             'user' => getenv('LHM_DATABASE_USER') ?: 'root',
             'pass' => getenv('LHM_DATABASE_PASSWORD') ?: null,
-            'port' => getenv('LHM_DATABASE_PORT') ?: 3306
+            'port' => getenv('LHM_DATABASE_PORT') ?: 3306,
         ];
 
         $this->adapter = new MysqlAdapter($options, null);
