@@ -1,15 +1,15 @@
 <?php
+
 namespace Lhm\Tests\Persistence;
 
 use Lhm\Lhm;
 use Lhm\Tests\Persistence\Migrations\HybridPhinxMigration;
+use Lhm\Tests\Persistence\Migrations\IndexMigration;
 use Lhm\Tests\Persistence\Migrations\InitialMigration;
 use Lhm\Tests\Persistence\Migrations\LhmMigration;
-use Phinx\Db\Table;
+use Lhm\Tests\Persistence\Migrations\RenameMigration;
 use Phinx\Migration\Manager\Environment;
 use Phinx\Migration\MigrationInterface;
-use Lhm\Tests\Persistence\Migrations\IndexMigration;
-use Lhm\Tests\Persistence\Migrations\RenameMigration;
 
 class LhmTest extends AbstractPersistenceTest
 {
@@ -17,7 +17,7 @@ class LhmTest extends AbstractPersistenceTest
     /** @var Environment */
     protected $environment;
 
-    protected function setUp()
+    protected function setUp(): void
     {
         parent::setUp();
 
@@ -25,7 +25,7 @@ class LhmTest extends AbstractPersistenceTest
         $this->environment->setAdapter($this->adapter);
     }
 
-    protected function tearDown()
+    protected function tearDown(): void
     {
         unset($this->environment, $this->migration);
         parent::tearDown();
@@ -34,14 +34,14 @@ class LhmTest extends AbstractPersistenceTest
     public function testMigrateLhm()
     {
         $time = time();
-        $this->environment->executeMigration(new InitialMigration($time - 1), MigrationInterface::UP);
-        $this->environment->executeMigration(new LhmMigration($time), MigrationInterface::UP);
+        $this->environment->executeMigration(new InitialMigration('test', $time - 1), MigrationInterface::UP);
+        $this->environment->executeMigration(new LhmMigration('test', $time), MigrationInterface::UP);
 
         $count = $this->adapter->query("SELECT COUNT(*) FROM ponies")->fetchColumn(0);
 
         /** @var \PDOStatement $statement */
         $statement = $this->adapter->query("SELECT nickname FROM ponies");
-        $nickname = $statement->fetchColumn(0);
+        $nickname  = $statement->fetchColumn(0);
 
         $this->assertEquals(100, $count);
         $this->assertEquals('derp', $nickname);
@@ -50,13 +50,13 @@ class LhmTest extends AbstractPersistenceTest
     public function testMigrateHybrid()
     {
         $time = time();
-        $this->environment->executeMigration(new InitialMigration($time - 1), MigrationInterface::UP);
-        $this->environment->executeMigration(new HybridPhinxMigration($time), MigrationInterface::UP);
+        $this->environment->executeMigration(new InitialMigration('test', $time - 1), MigrationInterface::UP);
+        $this->environment->executeMigration(new HybridPhinxMigration('test', $time), MigrationInterface::UP);
 
         /** @var \PDOStatement $statement */
         $statement = $this->adapter->query("SELECT age, location FROM ponies");
-        $age = $statement->fetchColumn(0);
-        $location = $statement->fetchColumn(1);
+        $age       = $statement->fetchColumn(0);
+        $location  = $statement->fetchColumn(1);
 
         $this->assertEquals(null, $age);
         $this->assertEquals('Canada', $location);
@@ -66,12 +66,17 @@ class LhmTest extends AbstractPersistenceTest
     {
         $this->adapter->query('CREATE TABLE test(id INT PRIMARY KEY);');
         $this->adapter->query('CREATE TABLE lhma_2015_03_04_13_22_33_test(id INT PRIMARY KEY);');
-        $this->adapter->query(implode("\n ", [
-            'CREATE TRIGGER lhmt_update_test',
-            'AFTER UPDATE ON test FOR EACH ROW',
-            'REPLACE INTO derp (`id`) /* large hadron migration (php) */',
-            'VALUES (NEW.`id`)'
-        ]));
+        $this->adapter->query(
+            implode(
+                "\n ",
+                [
+                    'CREATE TRIGGER lhmt_update_test',
+                    'AFTER UPDATE ON test FOR EACH ROW',
+                    'REPLACE INTO derp (`id`) /* large hadron migration (php) */',
+                    'VALUES (NEW.`id`)',
+                ]
+            )
+        );
 
         Lhm::setAdapter($this->adapter);
         Lhm::cleanup(true, ['until' => new \DateTime()]);
@@ -90,12 +95,17 @@ class LhmTest extends AbstractPersistenceTest
     {
         $this->adapter->query('CREATE TABLE test(id INT PRIMARY KEY);');
         $this->adapter->query('CREATE TABLE lhma_2015_03_04_13_22_33_test(id INT PRIMARY KEY);');
-        $this->adapter->query(implode("\n ", [
-            'CREATE TRIGGER lhmt_update_test',
-            'AFTER UPDATE ON test FOR EACH ROW',
-            'REPLACE INTO derp (`id`) /* large hadron migration (php) */',
-            'VALUES (NEW.`id`)'
-        ]));
+        $this->adapter->query(
+            implode(
+                "\n ",
+                [
+                    'CREATE TRIGGER lhmt_update_test',
+                    'AFTER UPDATE ON test FOR EACH ROW',
+                    'REPLACE INTO derp (`id`) /* large hadron migration (php) */',
+                    'VALUES (NEW.`id`)',
+                ]
+            )
+        );
 
         Lhm::setAdapter($this->adapter);
         Lhm::cleanup(false, ['until' => new \DateTime()]);
@@ -114,11 +124,11 @@ class LhmTest extends AbstractPersistenceTest
     public function testAddIndex()
     {
         $time = time();
-        $this->environment->executeMigration(new InitialMigration($time - 1), MigrationInterface::UP);
-        $this->environment->executeMigration(new IndexMigration($time), MigrationInterface::UP);
+        $this->environment->executeMigration(new InitialMigration('test', $time - 1), MigrationInterface::UP);
+        $this->environment->executeMigration(new IndexMigration('test', $time), MigrationInterface::UP);
 
         $exists = false;
-        $rows = $this->adapter->fetchAll(sprintf('SHOW INDEXES FROM %s', $this->adapter->quoteTableName('ponies')));
+        $rows   = $this->adapter->fetchAll(sprintf('SHOW INDEXES FROM %s', $this->adapter->quoteTableName('ponies')));
         foreach ($rows as $row) {
             if ($row['Key_name'] === 'ponies_age_idx') {
                 if ($row['Column_name'] === 'age' && $row['Non_unique'] === '1') {
@@ -134,9 +144,9 @@ class LhmTest extends AbstractPersistenceTest
     public function testRemoveIndex()
     {
         $time = time();
-        $this->environment->executeMigration(new InitialMigration($time - 1), MigrationInterface::UP);
-        $this->environment->executeMigration(new IndexMigration($time), MigrationInterface::UP);
-        $this->environment->executeMigration(new IndexMigration($time), MigrationInterface::DOWN);
+        $this->environment->executeMigration(new InitialMigration('test', $time - 1), MigrationInterface::UP);
+        $this->environment->executeMigration(new IndexMigration('test', $time), MigrationInterface::UP);
+        $this->environment->executeMigration(new IndexMigration('test', $time), MigrationInterface::DOWN);
 
         $rows = $this->adapter->fetchAll(sprintf('SHOW INDEXES FROM %s', $this->adapter->quoteTableName('ponies')));
         foreach ($rows as $row) {
@@ -147,11 +157,11 @@ class LhmTest extends AbstractPersistenceTest
     public function testRenameColumn()
     {
         $time = time();
-        $this->environment->executeMigration(new InitialMigration($time - 1), MigrationInterface::UP);
+        $this->environment->executeMigration(new InitialMigration('test', $time - 1), MigrationInterface::UP);
 
         $this->assertNotEmpty($this->adapter->fetchAll("SELECT * FROM ponies WHERE name IS NOT NULL"));
 
-        $this->environment->executeMigration(new RenameMigration($time), MigrationInterface::UP);
+        $this->environment->executeMigration(new RenameMigration('test', $time), MigrationInterface::UP);
 
         $this->assertEmpty($this->adapter->fetchAll("SELECT * FROM ponies WHERE first_name IS NULL"));
     }
