@@ -1,4 +1,5 @@
 <?php
+
 namespace Lhm\Tests\Persistence\Migrations;
 
 use Lhm\Lhm;
@@ -11,22 +12,29 @@ class IndexMigration extends AbstractMigration
     public function up()
     {
         Lhm::setAdapter($this->getAdapter());
-        Lhm::changeTable('ponies', function (Table $ponies) {
-            $ponies
-                ->addColumn('age', 'integer', ['null' => true])
-                ->addIndex('age', ['name' => 'ponies_age_idx'])
-                ->save();
-        });
+        Lhm::changeTable(
+            'ponies',
+            function (Table $ponies) {
+                $ponies
+                    ->addColumn('age', 'integer', ['null' => true])
+                    ->addIndex('age', ['name' => 'ponies_age_idx'])
+                    ->save();
+            }
+        );
     }
 
     public function down()
     {
         Lhm::setAdapter($this->getAdapter());
-        Lhm::changeTable('ponies', function (Table $ponies) {
-            $ponies
-                ->removeIndex('age', ['name' => 'ponies_age_idx'])
-                ->removeColumn('age', 'integer', ['null' => true])
-                ->save();
-        }, ['archive_name' => 'drop_the_ponies_age']);
+        Lhm::changeTable(
+            'ponies',
+            function (Table $ponies) {
+                $ponies
+                    ->removeIndex('age', ['name' => 'ponies_age_idx'])
+                    ->removeColumn('age', 'integer', ['null' => true])
+                    ->save();
+            },
+            ['archive_name' => 'drop_the_ponies_age']
+        );
     }
 }
