@@ -1,4 +1,5 @@
 <?php
+
 namespace Lhm\Tests\Persistence\Migrations;
 
 use Lhm\Lhm;
@@ -10,11 +11,18 @@ class LhmMigration extends AbstractMigration
     public function up()
     {
         Lhm::setAdapter($this->getAdapter());
-        Lhm::changeTable('ponies', function (Table $ponies) {
-            $ponies->addColumn('age', 'integer', ['default' => 1]);
-            $ponies->addColumn('nickname', 'string', ['after' => 'name', 'length' => 20, 'null' => true, 'default' => 'derp']);
-            $ponies->save();
-        });
+        Lhm::changeTable(
+            'ponies',
+            function (Table $ponies) {
+                $ponies->addColumn('age', 'integer', ['default' => 1]);
+                $ponies->addColumn(
+                    'nickname',
+                    'string',
+                    ['after' => 'name', 'length' => 20, 'null' => true, 'default' => 'derp']
+                );
+                $ponies->save();
+            }
+        );
     }
 
     public function down()
