@@ -1,4 +1,5 @@
 <?php
+
 namespace Lhm\Tests\Persistence\Migrations;
 
 use Lhm\Lhm;
@@ -10,10 +11,13 @@ class RenameMigration extends AbstractMigration
     public function up()
     {
         Lhm::setAdapter($this->getAdapter());
-        Lhm::changeTable('ponies', function (Table $ponies) {
-            $ponies
-                ->renameColumn('name', 'first_name')
-                ->save();
-        });
+        Lhm::changeTable(
+            'ponies',
+            function (Table $ponies) {
+                $ponies
+                    ->renameColumn('name', 'first_name')
+                    ->save();
+            }
+        );
     }
 }
